hello: This is the first commit to the pyxen repo
This commit adds the directory structure to be followed during
the first stage of the project. The benchmarking of the rpython
code run on top of the Xen Hypervisor, since at the moment this
is the only approach that is in our grasp for pyxen. If the benchmarks
prove to be promising, we can proceed with the second stage,
cross-compiling CPython in a single binary blob to run the whole
python environment in xen StubDom

Signed-off-by: Razvan Ghitulete <razvan.ghitulete@gmail.com>